requirements.txt
torch==1.10.2
numpy==1.19.5
torchvision==0.11.3
pandas==1.1.5
fastapi==0.78.0
Pillow==9.2.0
scikit_learn==0.23.0
